Add common generator definitions
