+++ src/ImageGenerators/MomentGenerator.cc
@@ -6,16 +6,12 @@
 
 #include "MomentGenerator.h"
 
+#include "ImageGenerator.h"
 #include "../Logger/Logger.h"
 
 using namespace carta;
 
 using IM = ImageMoments<casacore::Float>;
-
-static const int FIRST_PROGRESS_AFTER_MILLI_SECS = 5000;
-static const float PROGRESS_REPORT_INTERVAL = 0.1;
-static const float PROCESS_COMPLETED = 1;
-static const int ID_MULTIPLIER = 1000;
 
 MomentGenerator::MomentGenerator(const casacore::String& filename, casacore::ImageInterface<float>* image)
     : _filename(filename), _image(image), _sub_image(nullptr), _image_moments(nullptr), _success(false), _cancel(false) {
@@ -23,7 +19,7 @@ MomentGenerator::MomentGenerator(const casacore::String& filename, casacore::Ima
 }
 
 bool MomentGenerator::CalculateMoments(int file_id, const casacore::ImageRegion& image_region, int spectral_axis, int stokes_axis,
-    const MomentProgressCallback& progress_callback, const CARTA::MomentRequest& moment_request, CARTA::MomentResponse& moment_response,
+    const GeneratorProgressCallback& progress_callback, const CARTA::MomentRequest& moment_request, CARTA::MomentResponse& moment_response,
     std::vector<CollapseResult>& collapse_results) {
     _spectral_axis = spectral_axis;
     _stokes_axis = stokes_axis;
